Move Step and Procedure into single folder
