+++ src/components/ComponentFactory.js
@@ -48,7 +48,7 @@ import MongoWebShell from './MongoWebShell';
 import OpenAPI from './OpenAPI';
 import Operation from './Operation';
 import Paragraph from './Paragraph';
-import Procedure from './Procedure';
+import Procedure from './Procedure/Procedure';
 import QuizChoice from './Widgets/QuizWidget/QuizChoice';
 import QuizWidget from './Widgets/QuizWidget/QuizWidget';
 import Reference from './Reference';
@@ -58,7 +58,6 @@ import Root from './Root';
 import Rubric from './Rubric';
 import SearchResults from './SearchResults';
 import Section from './Section';
-import Step from './Step';
 import Strong from './Strong';
 import Subscript from './Subscript';
 import SubstitutionReference from './SubstitutionReference';
@@ -187,7 +186,6 @@ const componentMap = {
   'search-results': SearchResults,
   section: Section,
   sharedinclude: Include,
-  step: Step,
   strong: Strong,
   substitution_reference: SubstitutionReference,
   tabs: Tabs,
