+++ tests/unit/Procedure.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render } from '@testing-library/react';
-import Procedure from '../../src/components/Procedure';
+import Procedure from '../../src/components/Procedure/Procedure';
 
 // data for this component
 import mockData from './data/Procedure.test.json';
