+++ tests/unit/Step.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render } from '@testing-library/react';
-import Step from '../../src/components/Step';
+import Step from '../../src/components/Procedure/Step';
 
 // data for this component
 import mockData from './data/Step.test.json';
